requirements.txt
# Essential Python Libraries
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1

# Django and Web-related Libraries
Django==4.2.5
django-google-translate==1.1
django-phonenumber-field==7.1.0
sqlparse==0.4.4

# Google AI and APIs
google-ai-generativelanguage==0.3.2
google-api-core==2.12.0
google-auth==2.23.1
google-generativeai==0.2.0
googleapis-common-protos==1.60.0

# Data Science & Machine Learning (Required for .pkl models)
joblib==1.3.2
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
scipy==1.11.2
threadpoolctl==3.2.0

# Utility Libraries
phonenumberslite==8.13.21
Pillow==10.0.1
prometheus-client==0.17.1
protobuf==4.24.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
